Turns missing pr in workflow_run into warning. (#5)
* Turns missing pr in workflow_run into warning.

Sometimes (quite often really) when PR gets approved, the PR
gets merged rather quickly, without waiting for result of this
action. Or a new PR gets pushed quickly. In those cases PR will
not be found. But this is usually not a problem then and rather
than failing, we should simply print a warning and exit.

* fixup! Turns missing pr in workflow_run into warning.

Co-authored-by: Tobiasz Kędzierski <tobiasz.kedzierski@polidea.com>