Add `remove_label_when_approval_missing` input. (#6)
Removing label when you lost or never had an approval status
might not be what you desire to do. In some workflows when
the label is manually set, removing it is automatically is
harmful and it's better to leave the decision about removal
to the commiters.

Adding the parameter solves the problem.